feat: Losses class we developed is renamed and adjusted to single_link_losses. Now single_link_redundancy run is included. Previously developed single_link_losses is renamed to _single_link_losses
