+++ examples/data/losses/analysis_losses.ini
@@ -3,7 +3,8 @@ name = republic_dominican
 
 [analysis1]
 name = republic_dominican_redundancy
-analysis = losses
+analysis = single_link_losses
+weighing = time
 duration_event = 600
 part_of_day = day
 performance_metric = diff_time
