+++ ra2ce/analysis/analysis_config_data/analysis_config_data.py
@@ -86,11 +86,6 @@ class AnalysisSectionIndirect(AnalysisSectionBase):
     traffic_cols: list[str] = field(default_factory=list)
     # losses
     duration_event: float = math.nan # TODO remove the deprecated attribute that have been replaced by csv
-    duration_disruption: float = math.nan
-    fraction_detour: float = math.nan
-    fraction_drivethrough: float = 0.0
-    rest_capacity: float = math.nan
-    maximum_jam: float = math.nan
     part_of_day: PartOfDayEnum = field(default_factory=lambda: PartOfDayEnum.DAY)
     performance_metric: str = "diff_time"  # "diff_time" or "diff_length" relates to the used criticality metric
     resilience_curve_file: Optional[Path] = None
