+++ ra2ce/analysis/analysis_factory.py
@@ -160,15 +160,11 @@ class AnalysisFactory:
                 origins_destinations=analysis_config.config_data.origins_destinations,
                 file_id=analysis_config.config_data.network.file_id,
             )
-        if analysis.analysis == AnalysisIndirectEnum.LOSSES:
-            return Losses(
+        if analysis.analysis == AnalysisIndirectEnum.SINGLE_LINK_LOSSES:
+            return SingleLinkLosses(
                 network=analysis_config.config_data.network,
                 graph_file=analysis_config.graph_files.base_graph_hazard,
                 **_input_dict
-            )
-        if analysis.analysis == AnalysisIndirectEnum.SINGLE_LINK_LOSSES:
-            return SingleLinkLosses(
-                graph_file=analysis_config.graph_files.base_graph_hazard, **_input_dict
             )
         if analysis.analysis == AnalysisIndirectEnum.MULTI_LINK_LOSSES:
             return MultiLinkLosses(
